Created SpecialDay record
