+++ src/main/java/com/example/mealplanadmin/model/Plan.java
@@ -8,8 +8,8 @@ public record Plan(Long id, LocalDate startDate, Integer totalDays, Integer meal
         this(null, startDate, totalDays, mealsPerDay);
     }
 
-    public static Plan fromEntity(PlanEntity planEntity) {
-        return new Plan(planEntity.getId(), planEntity.getStartDate(), planEntity.getTotalDays(), planEntity.getMealsPerDay());
+    public Plan(PlanEntity planEntity) {
+        this(planEntity.getId(), planEntity.getStartDate(), planEntity.getTotalDays(), planEntity.getMealsPerDay());
     }
 
     public PlanEntity toEntity() {
